Extract pubsub.go
